place all integration tests in one file
